add roles system

+++ src/main/java/com/example/adminpanelbackend/dataBase/service/AdminService.java
@@ -3,5 +3,5 @@ package com.example.adminpanelbackend.dataBase.service;
 import com.example.adminpanelbackend.dataBase.entity.AdminEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminService extends JpaRepository<AdminEntity, Integer> {
+public interface AdminService extends JpaRepository<AdminEntity, Long> {
 }
